Move fake data utils to own module
